add no-outside-click to sql editor in dashboard eidit
