+++ frontend/src/metabase/components/Modal/MaybeOnClickOutsideWrapper.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import OnClickOutsideWrapper from "metabase/components/OnClickOutsideWrapper";
+
+export function MaybeOnClickOutsideWrapper({
+  children,
+  noOnClickOutsideWrapper = false,
+  ...props
+}: {
+  children: React.ReactNode;
+  noOnClickOutsideWrapper?: boolean;
+} & React.ComponentProps<typeof OnClickOutsideWrapper>) {
+  return noOnClickOutsideWrapper ? (
+    <div>{children}</div>
+  ) : (
+    <OnClickOutsideWrapper {...props}>{children}</OnClickOutsideWrapper>
+  );
+}
