+++ frontend/src/metabase/components/Modal/Modal.tsx
@@ -1,9 +1,11 @@
-/* eslint-disable react/prop-types */
 import React from "react";
 import { RoutelessFullPageModal } from "metabase/components/Modal/RoutelessFullPageModal";
 import { WindowModal } from "metabase/components/Modal/WindowModal";
 
-const Modal = ({ full = false, ...props }) =>
+const Modal = ({
+  full = false,
+  ...props
+}: { full: boolean; isOpen: boolean } & { [_: string]: any }) =>
   full ? (
     props.isOpen ? (
       <RoutelessFullPageModal {...props} />
